override the tick method entirely
this should give us more freedom to change the projectile's ticking behaviour.  note whilst the code is copypasted from mojang's code, it will be modified a lot.

the first thing done was a fix particle emission logic and mathematics.

secondly the calculations surrounding water particles was simplified.

the check for fire was removed so that blaze bullets don't make themselves obvious.