fix usage of ammo stack counts and tiers
now guns pick ammo based on the following rules in this order:

pick a stack that has at least the minimum cost of the weapon
pick the most powerful damage
pick the leftmost stack

this means a player with a 2 cost weapon cannot use a 1 blaze stack, but could use a 2 flint stack unless a hunger or xp bullet exists.

hunger bullets ignore the requirement for the player to have the minimum cost of the weapon, allowing for a suicide attack.